scripts/requirements.txt
openai>=1.30.0
feedparser>=6.0.0
playwright>=1.44.0
python-dotenv>=1.0.0
requests>=2.31.0
